[dtensor] enable foreach operators for adam optimizer
This PR enables basic foreach ops in DTensor for adam optimizer, to improve performance
compare to optimizer using torch.Tensor. Currently by default optimizer
won't do this for tensor subclass, we will need to enable this by
default in DTensor when all ops are covered, or early enable this when
exploring new FSDP, we just need to append DTensor to the optimizer
allow list.

cc awgu

ghstack-source-id: b37c497620b9b15895bbe997c4774e601b2bf9fd
Pull Request resolved: https://github.com/pytorch/pytorch/pull/112108